requirements.txt
python-dotenv
langchain
openai
panel
langchain-community
chromadb
pypdf
docarray
tiktoken